Ensuring paths for test inputs

+++ tests/test_api.py
@@ -17,6 +17,11 @@ class TestNewApi(unittest.TestCase):
         cls.path_to_file = Path.cwd().joinpath("tests", "input", "test_resource_fr.po")
         cls.output_dir = Path.cwd().joinpath("tests", "output")
 
+        if missing := next(filter(lambda p: not p.exists(), [cls.path_to_file]), None):
+            raise ValueError(
+                f"Unable to complete test with broken tests inputs. Found missing: {missing}"
+            )
+
         if project := cls.tx.get_project(project_slug=cls.project_slug):
             print("Found old project, removing.")
             project.delete()
